Remove TypeScript project warning from project creation
The confirmation prompt and warning message regarding the use of TypeScript in 'create_new_project.ts' have been removed. As the TypeScript Skeleton project is no longer under development and can be used directly, the contingency to switch to JavaScript is unnecessary. This streamlines the process of creating a new project.